requirements.txt
fastapi==0.115.4
uvicorn==0.32.0
python-dotenv==1.0.1
openai>=1.47.0
requests==2.32.3
pydub==0.25.1
sqlalchemy==2.0.36
pydantic==2.9.2
pydantic-settings==2.6.1
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
aiofiles==24.1.0